pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stackhawk-mcp"
version = "1.0.0"
description = "StackHawk MCP Server for Security Analytics and Developer Integration"
authors = [{name = "StackHawk, Inc.", email = "support@stackhawk.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0",
    "jsonschema>=4.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0"
]

[project.scripts]
stackhawk-mcp = "stackhawk_mcp.server:main"

[project.urls]
Homepage = "https://github.com/stackhawk/stackhawk-mcp"
Repository = "https://github.com/stackhawk/stackhawk-mcp"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]